Fix broken intra-page links in the header navbar
The link to the “Papers” section, was using an out-of-date (I presume) anchor (`#papers` instead of the current `#papers-preprints-and-software`).

The “Talks” link was erroneously using whitespaces instead of hyphens (`#selected talks and posters` instead of `#selected-talks-and-posters`).

This commit fixes that. 
All the header navigation links should be working now. 
I haven’t tested, tbh. Changes seem safe enough I didn’t bother deploying a local copy of the site.